Code reveiw fixes
