+++ tests/components/huum/test_config_flow.py
@@ -2,12 +2,11 @@
 
 from unittest.mock import AsyncMock, patch
 
-from huum.exceptions import Forbidden, NotAuthenticated
+from huum.exceptions import Forbidden
 import pytest
 
 from homeassistant import config_entries
 from homeassistant.components.huum.const import DOMAIN
-from homeassistant.config_entries import ConfigEntryState
 from homeassistant.const import CONF_PASSWORD, CONF_USERNAME
 from homeassistant.core import HomeAssistant
 from homeassistant.data_entry_flow import FlowResultType
@@ -140,6 +139,7 @@ async def test_reauth_flow(
 
     assert result["type"] is FlowResultType.ABORT
     assert result["reason"] == "reauth_successful"
+    assert mock_config_entry.data[CONF_USERNAME] == TEST_USERNAME
     assert mock_config_entry.data[CONF_PASSWORD] == "new_password"
 
 
@@ -186,26 +186,5 @@ async def test_reauth_errors(
 
     assert result["type"] is FlowResultType.ABORT
     assert result["reason"] == "reauth_successful"
+    assert mock_config_entry.data[CONF_USERNAME] == TEST_USERNAME
     assert mock_config_entry.data[CONF_PASSWORD] == "new_password"
-
-
-@pytest.mark.parametrize("side_effect", [Forbidden, NotAuthenticated])
-async def test_auth_error_triggers_reauth(
-    hass: HomeAssistant,
-    mock_config_entry: MockConfigEntry,
-    side_effect: type[Exception],
-) -> None:
-    """Test that an auth error during coordinator refresh triggers reauth."""
-    mock_config_entry.add_to_hass(hass)
-
-    with patch(
-        "homeassistant.components.huum.coordinator.Huum.status",
-        side_effect=side_effect,
-    ):
-        await hass.config_entries.async_setup(mock_config_entry.entry_id)
-        await hass.async_block_till_done()
-
-    assert mock_config_entry.state is ConfigEntryState.SETUP_ERROR
-    assert any(
-        mock_config_entry.async_get_active_flows(hass, {config_entries.SOURCE_REAUTH})
-    )
